+++ CHANGELOG.md
@@ -83,6 +83,7 @@ Semantic versioning in our case means:
   for `WPS211` complexity check anymore, #2394
 - `WPS303`: Underscore (`_`) now only allowed with 3 digits after it, #3120
 - Adds new rule to forbid `lambda` assigns to special attributes, #1733
+- Adds new rule to check problematic function params, #1343
 - Adds support to run `wemake-python-styleguide` as a `pre-commit` hook, #2588
 
 ### Bugfixes